progress, bug in mtd-f WIP

+++ include/ZobristTable.hpp
@@ -16,8 +16,8 @@ namespace ZobristTable {
     std::array<std::array<uint64_t, S>, SIZE>   _init_zobrist_table(void) {
         std::array<std::array<uint64_t, S>, SIZE> table;
         std::srand(std::time(nullptr));
-        for (uint16_t n = 0; n < SIZE; n++) {
-            for (uint16_t s = 0; s < S; s++)
+        for (int n = 0; n < SIZE; n++) {
+            for (int s = 0; s < S; s++)
                 table[n][s] = std::rand();
         }
         return (table);
@@ -36,14 +36,14 @@ namespace ZobristTable {
     struct KeyHash {
         uint64_t operator()(const Key &k) const {
             uint64_t    hash = 0;
-            for (uint16_t n = 0; n < SIZE; n++)
+            for (int n = 0; n < SIZE; n++)
                 hash ^= _table[n][(k.p1[n] == 0 ? (k.p2[n] == 0 ? 0 : 2) : 1)]; // could use a 3rd dimension to avoid branching (opti?)
                 // hash ^= _table[n][(k.p1[n]<<(n&0x3F))^(k.p2[n]<<(n&0x3F-1))>>0x3F]; // maybe faster ?
                 // hash ^= _table[n][k.p1[n]][k.p2[n]]; // could use a 3rd dimension to avoid branching (opti?)
             return (hash);
         }
     };
-    std::unordered_map<Key, uint16_t, KeyHash>  map;
+    std::unordered_map<Key, int32_t, KeyHash>  map;
 }
 
 #endif
